+++ tests/DetectsChangesTest.php
@@ -37,7 +37,7 @@ class DetectsChangesTest extends TestCase
             ],
         ];
 
-        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes->toArray());
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
     }
 
     /** @test */
@@ -76,7 +76,7 @@ class DetectsChangesTest extends TestCase
             ],
         ];
 
-        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes->toArray());
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
     }
 
     /** @test */
@@ -100,7 +100,7 @@ class DetectsChangesTest extends TestCase
             ],
         ];
 
-        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes->toArray());
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
     }
 
     /** @test */
@@ -121,7 +121,7 @@ class DetectsChangesTest extends TestCase
             ],
         ];
 
-        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes->toArray());
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
     }
 
     /** @test */
@@ -148,7 +148,7 @@ class DetectsChangesTest extends TestCase
             ],
         ];
 
-        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes->toArray());
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
     }
 
     /** @test */
@@ -189,7 +189,7 @@ class DetectsChangesTest extends TestCase
                 ],
         ];
 
-        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes->toArray());
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
     }
 
     /** @test */
@@ -228,7 +228,7 @@ class DetectsChangesTest extends TestCase
                 ],
         ];
 
-        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes->toArray());
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
     }
 
     /** @test */
@@ -246,7 +246,7 @@ class DetectsChangesTest extends TestCase
 
         $article->save();
 
-        $this->assertEquals(collect(), $this->getLastActivity()->changes);
+        $this->assertEquals(collect(), $this->getLastActivity()->changes());
     }
 
     /** @test */
@@ -263,7 +263,7 @@ class DetectsChangesTest extends TestCase
         ]);
 
         $this->assertEquals('deleted', $this->getLastActivity()->description);
-        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes);
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes());
     }
 
     /** @test */
@@ -296,7 +296,57 @@ class DetectsChangesTest extends TestCase
                 ],
             ],
         ];
-        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes->toArray());
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
+    }
+
+    /** @test */
+    public function it_can_use_fillable_as_loggable_attributes()
+    {
+        $articleClass = new class() extends Article {
+            protected $fillable = ['name', 'text'];
+            protected static $logFillable = true;
+
+            use LogsActivity;
+        };
+
+        $article = new $articleClass();
+        $article->name = 'my name';
+        $article->save();
+
+        $expectedChanges = [
+            'attributes' => [
+                'name' => 'my name',
+                'text' => null,
+            ],
+        ];
+
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
+    }
+
+    /** @test */
+    public function it_can_use_both_fillable_and_log_attributes()
+    {
+        $articleClass = new class() extends Article {
+            protected $fillable = ['name'];
+            protected static $logAttributes = ['text'];
+            protected static $logFillable = true;
+
+            use LogsActivity;
+        };
+
+        $article = new $articleClass();
+        $article->name = 'my name';
+        $article->text = 'my text';
+        $article->save();
+
+        $expectedChanges = [
+            'attributes' => [
+                'name' => 'my name',
+                'text' => 'my text',
+            ],
+        ];
+
+        $this->assertEquals($expectedChanges, $this->getLastActivity()->changes()->toArray());
     }
 
     protected function createArticle(): Article