+++ tests/Models/CustomActivityModel.php
@@ -6,4 +6,8 @@ use RifkyEkayama\Activitylog\Models\Activity;
 
 class CustomActivityModel extends Activity
 {
+    public function getCustomPropertyAttribute()
+    {
+        return $this->changes();
+    }
 }